perf(irodori): codec を CUDA bf16 に移してRTF 5.65x → 0.27x へ
プロファイルした結果 decode_latent (DACVAE codec の decode) が 16.6 秒と
支配的だった (fp32/codec CPU 時)。codec を CUDA に移し、model と codec を
両方 bf16 に揃えることで dtype mismatch も同時解消。

計測 (RTX 4080, num_steps=16, warm, 音声 3.6秒):
- fp32 + codec CPU:      19s  / RTF 5.65x  ← 旧既定
- fp32 + codec CUDA:     1.8s / RTF 0.50x
- bf16 + codec CUDA bf16: 0.9s / RTF 0.27x  ← 新既定

副次修正:
- SamplingResult.audio が bf16 の場合 np.asarray が "Got unsupported ScalarType"
  で落ちるため、float32 へ明示キャストしてから numpy 化
- config/default.json の engines.irodori 既定値を bf16/CUDA codec に更新

参考 (同じ文で num_steps を振った場合):
- steps= 8: 0.63s / RTF 0.17x
- steps=16: 0.94s / RTF 0.26x (品質と速度のバランス、新既定の registry 想定)
- steps=20: 1.07s / RTF 0.29x (上流既定に近い)

+++ tools/speak/engine/irodori.py
@@ -160,10 +160,12 @@ class IrodoriEngine(TTSEngine):
             )
         )
 
-        # SamplingResult.audio は torch.Tensor (shape: [samples] or [channels, samples])
+        # SamplingResult.audio は torch.Tensor (shape: [samples] or [channels, samples])。
+        # bf16 運用時は numpy が BFloat16 を扱えないので、明示的に float32 にキャスト。
+        import torch  # local import: lazy_load 経由で既に存在する
         audio_tensor = result.audio
         if hasattr(audio_tensor, "detach"):
-            audio_tensor = audio_tensor.detach().cpu()
+            audio_tensor = audio_tensor.detach().to(torch.float32).cpu()
         audio = np.asarray(audio_tensor, dtype=np.float32)
         if audio.ndim > 1:
             # ステレオは平均して mono 化 (playback_worker は mono 前提)
